+++ designsafe/static/scripts/data-depot/components/projects/curation-directory/curation-directory.component.js
@@ -91,15 +91,15 @@ class CurationDirectoryCtrl {
 
     goPreview() {
         if (this.browser.project.value.projectType === 'experimental') {
-            this.$state.go('projects.preview', {projectId: this.browser.project.uuid, data: this.browser});
+            this.$state.go('projects.preview', {projectId: this.browser.project.uuid});
         } else if (this.browser.project.value.projectType === 'simulation') {
-            this.$state.go('projects.previewSim', {projectId: this.browser.project.uuid, data: this.browser});
+            this.$state.go('projects.previewSim', {projectId: this.browser.project.uuid});
         } else if (this.browser.project.value.projectType === 'hybrid_simulation') {
-            this.$state.go('projects.previewHybSim', {projectId: this.browser.project.uuid, data: this.browser});
+            this.$state.go('projects.previewHybSim', {projectId: this.browser.project.uuid});
         } else if (this.browser.project.value.projectType === 'other') {
-            this.$state.go('projects.previewOther', {projectId: this.browser.project.uuid, data: this.browser});
+            this.$state.go('projects.previewOther', {projectId: this.browser.project.uuid});
         } else if (this.browser.project.value.projectType === 'field_recon') {
-            this.$state.go('projects.previewFieldRecon', {projectId: this.browser.project.uuid, data: this.browser});
+            this.$state.go('projects.previewFieldRecon', {projectId: this.browser.project.uuid});
         }
     }