+++ designsafe/static/scripts/data-depot/components/projects/project-view/project-view.component.js
@@ -81,16 +81,13 @@ class ProjectViewCtrl {
     this.ProjectService.editProject(this.browser.project);
   }
 
-  manageProjectType($event) {
-    if ($event) {
-      $event.preventDefault();
-    }
+  overview() {
     this.$uibModal.open({
-      component: 'manageProjectType',
-      resolve: {
-        options: () => { return { 'project': this.browser.project, 'warning': false }; },
-      },
-      size: 'lg',
+        component: 'manageProjectType',
+        resolve: {
+            options: () => { return {'project': this.browser.project, 'preview': true, 'warning': false}; },
+        },
+        size: 'lg',
     });
   }
 
@@ -109,25 +106,15 @@ class ProjectViewCtrl {
 
   publicationPreview() {
     if (this.browser.project.value.projectType === 'experimental') {
-      this.$state.go('projects.preview', { projectId: this.browser.project.uuid, data: this.browser}).then(() => {
-        this.checkState();
-      });
+      this.$state.go('projects.preview', { projectId: this.browser.project.uuid});
     } else if (this.browser.project.value.projectType === 'simulation') {
-      this.$state.go('projects.previewSim', { projectId: this.browser.project.uuid, data: this.browser}).then(() => {
-        this.checkState();
-      });
+      this.$state.go('projects.previewSim', { projectId: this.browser.project.uuid});
     } else if (this.browser.project.value.projectType === 'hybrid_simulation') {
-      this.$state.go('projects.previewHybSim', { projectId: this.browser.project.uuid, data: this.browser}).then(() => {
-        this.checkState();
-      });
+      this.$state.go('projects.previewHybSim', {projectId: this.browser.project.uuid});
     } else if (this.browser.project.value.projectType === 'other') {
-      this.$state.go('projects.previewOther', { projectId: this.browser.project.uuid, data: this.browser}).then(() => {
-        this.checkState();
-      });
+      this.$state.go('projects.previewOther', { projectId: this.browser.project.uuid});
     } else if (this.browser.project.value.projectType === 'field_recon') {
-      this.$state.go('projects.previewFieldRecon', { projectId: this.browser.project.uuid, data: this.browser}).then(() => {
-        this.checkState();
-      });
+      this.$state.go('projects.previewFieldRecon', { projectId: this.browser.project.uuid});
     } else {
       this.manageProjectType();
     }